requirements.txt
pandas
opencv-python
matplotlib
tqdm
scikit-learn
scipy
Pillow
ipython 
jupyter
bertopic
importlib-resources
transformers
scenedetect==0.6.1
timm
